Final_Project_mantler/requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
geopandas==0.14.4
rasterio==1.3.10
shapely==2.0.4
pyproj==3.6.1
numpy==1.26.4
pandas==2.2.2
requests==2.32.3
